refactor: move commands from internal to cmd
