+++ cmd/artifact/info/doc.go
