+++ cmd/index/add/doc.go
