+++ cmd/index/list/doc.go
