+++ cmd/index/remove/doc.go
