+++ cmd/index/remove/remove.go
