+++ cmd/registry/auth/auth.go
@@ -1,4 +1,4 @@
-// Copyright 2022 The Falco Authors
+// Copyright 2023 The Falco Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,31 +12,29 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package registry
+package auth
 
 import (
 	"context"
 
 	"github.com/spf13/cobra"
 
-	"github.com/falcosecurity/falcoctl/internal/registry/auth"
-	"github.com/falcosecurity/falcoctl/internal/registry/pull"
-	"github.com/falcosecurity/falcoctl/internal/registry/push"
+	"github.com/falcosecurity/falcoctl/cmd/registry/auth/basic"
+	"github.com/falcosecurity/falcoctl/cmd/registry/auth/oauth"
 	commonoptions "github.com/falcosecurity/falcoctl/pkg/options"
 )
 
-// NewRegistryCmd returns the registry command.
-func NewRegistryCmd(ctx context.Context, opt *commonoptions.CommonOptions) *cobra.Command {
+// NewAuthCmd returns the registry command.
+func NewAuthCmd(ctx context.Context, opt *commonoptions.CommonOptions) *cobra.Command {
 	cmd := &cobra.Command{
-		Use:                   "registry",
+		Use:                   "auth",
 		DisableFlagsInUseLine: true,
-		Short:                 "Interact with OCI registries",
-		Long:                  "Interact with OCI registries",
+		Short:                 "Handle authentication towards OCI registries",
+		Long:                  "Handle authentication towards OCI registries",
 	}
 
-	cmd.AddCommand(auth.NewAuthCmd(ctx, opt))
-	cmd.AddCommand(push.NewPushCmd(ctx, opt))
-	cmd.AddCommand(pull.NewPullCmd(ctx, opt))
+	cmd.AddCommand(basic.NewBasicCmd(ctx, opt))
+	cmd.AddCommand(oauth.NewOauthCmd(ctx, opt))
 
 	return cmd
 }
