+++ cmd/registry/auth/basic/basic.go
