+++ cmd/registry/auth/basic/doc.go
