+++ cmd/registry/auth/doc.go
