+++ cmd/registry/auth/oauth/doc.go
