+++ cmd/registry/auth/oauth/oauth.go
