+++ cmd/registry/doc.go
