+++ cmd/registry/pull/doc.go
