+++ cmd/registry/pull/pull.go
