+++ cmd/registry/push/doc.go
