+++ cmd/version/doc.go
